Added Walls And A Character
Walls are in and they fade away based on the camera's angle. They're
annoying to set up because they aren't handles procedurally, but I'll
live. Only five levels to make it for. The character is in and works
with a navmesh. Next up is mechanics stuff. Still need to work out the
design side of that jazz...