Added Interactable Objects
Lever and button are in, and an interaction script is in. This doesn't
actually work yet, however. It makes the character walk to the correct
spot to use the object, but it still lacks the functionality to make the
interactable object activate. Still need a game manager, audio
player/manager, splash text, and level transitions. So much to do, so
very little time.